Add cloud storage page to settings navigation
