+++ Ink Canvas/Windows/SettingsViews/Pages/HomePage.xaml
@@ -153,6 +153,13 @@
                         </ui:SettingsCard.HeaderIcon>
                     </ui:SettingsCard>
 
+                    <ui:SettingsCard Header="云存储" Description="云存储与自动上传"
+                        IsClickEnabled="True" Click="QuickNavCard_Click" Tag="CloudStoragePage">
+                        <ui:SettingsCard.HeaderIcon>
+                            <ui:FontIcon Icon="{x:Static ui:SegoeFluentIcons.Cloud}" />
+                        </ui:SettingsCard.HeaderIcon>
+                    </ui:SettingsCard>
+
                     <ui:SettingsCard Header="随机点名" Description="随机点名与计时器"
                         IsClickEnabled="True" Click="QuickNavCard_Click" Tag="RandomDrawPage">
                         <ui:SettingsCard.HeaderIcon>
